# AArch64 BMLA: Use pairwise addition to accumulate popcount results