## What do these changes do?
This changes the accumulate of the popcount result in the AArch BMLA to use pairwise addition. See [Figure 7.9 in the arm docs](https://developer.arm.com/docs/den0024/latest/aarch64-floating-point-and-neon/aarch64-neon-instruction-format). The final accumulate is then done using [`uadalp`](https://developer.arm.com/docs/dui0801/j/a64-simd-vector-instructions/uadalp-vector).
This removes the need for the `ins` instructions, while keeping the same amount of arithmetic operations.
I couldn't find the cycle counts for these operations so I am not sure what the theoretical speedup of this should be, let me know if I am missing something here.

## How Has This Been Tested?
CI

## Benchmark Results
~~Initial benchmarks on my phone in the warm easter sun were promising, but I will properly benchmark the changes next week in a controlled environment without my phone overheating.~~

I ran some full model benchmarks on my phone with `num_calls=250`:

| Model          | Threads | master (ms)    | PR (ms)        |
| -------------- | ------- | -------------- | -------------- |
| QuickNet       | 1       | 25.2 ± 0.3     | **24.3 ± 0.3** |
| QuickNet       | 2       | 18.4 ± 0.3     | **17.5 ± 0.5** |
| QuickNet       | 4       | **15.0 ± 0.3** | 15.2 ± 0.4     |
|                |         |                |                |
| QuickNet Large | 1       | 36.8 ± 0.3     | **35.7 ± 0.3** |
| QuickNet Large | 2       | 26.4 ± 0.3     | **25.9 ± 0.3** |
| QuickNet Large | 4       | 21.8 ± 0.3     | **21.7 ± 0.5** |
|                |         |                |                |
| QuickNet XL    | 1       | 66.7 ± 0.3     | **64.2 ± 0.3** |
| QuickNet XL    | 2       | 47.1 ± 0.5     | **44.9 ± 0.4** |
| QuickNet XL    | 4       | 38.9 ± 0.3     | **37.9 ± 0.2** |