@lgeiger 2020-04-12
I updated the benchmark table above and on my phone this PR results in an performance improvement of **2.6%** on average (3.4 %, 3.8 % and 0.6 % for 1, 2 or 4 threaded execution respectively).