@lgeiger 2020-04-14
With this implementation we can also remove the need for one temporary register if we reorder the BMLA to:
```cpp
// temporary NEON registers: v28,v29,v30
#define LCE_BMLA(Vd, Vr, Vl1, Vl2, Vl3, Vl4) \
  "eor v28.16b, " #Vr".16b, " #Vl1".16b\n"    \
  "eor v29.16b, " #Vr".16b, " #Vl2".16b\n"    \
  "cnt v28.16b, v28.16b\n"                    \
  "cnt v29.16b, v29.16b\n"                    \
  "addp v28.16b, v28.16b, v29.16b\n"          \
  "eor v29.16b, " #Vr".16b, " #Vl3".16b\n"    \
  "eor v30.16b, " #Vr".16b, " #Vl4".16b\n"    \
  "cnt v29.16b, v29.16b\n"                    \
  "cnt v30.16b, v30.16b\n"                    \
  "addp v29.16b, v29.16b, v30.16b\n"          \
  "addp v28.16b, v28.16b, v29.16b\n"          \
  "uaddlp v28.8h, v28.16b\n"                  \
  "uadalp " #Vd".4s, v28.8h\n"
```
@larq/compute-engine do you think this would be useful?