@arashb 2020-04-14
> With this implementation we can also remove the need for one temporary register if we reorder the BMLA to:
> 
> ```c++
> // temporary NEON registers: v28,v29,v30
> #define LCE_BMLA(Vd, Vr, Vl1, Vl2, Vl3, Vl4) \
>   "eor v28.16b, " #Vr".16b, " #Vl1".16b\n"    \
>   "eor v29.16b, " #Vr".16b, " #Vl2".16b\n"    \
>   "cnt v28.16b, v28.16b\n"                    \
>   "cnt v29.16b, v29.16b\n"                    \
>   "addp v28.16b, v28.16b, v29.16b\n"          \
>   "eor v29.16b, " #Vr".16b, " #Vl3".16b\n"    \
>   "eor v30.16b, " #Vr".16b, " #Vl4".16b\n"    \
>   "cnt v29.16b, v29.16b\n"                    \
>   "cnt v30.16b, v30.16b\n"                    \
>   "addp v29.16b, v29.16b, v30.16b\n"          \
>   "addp v28.16b, v28.16b, v29.16b\n"          \
>   "uaddlp v28.8h, v28.16b\n"                  \
>   "uadalp " #Vd".4s, v28.8h\n"
> ```
> 
> @larq/compute-engine do you think this would be useful?

changing the order of instructions changes the data dependency of instructions and therefore the number of cycles for dual-issue instructions (see the difference the between **execution latency** and **throughput** [here](http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.100400_0001_03_en/wzz1471434459698.html)). Right now, we are also not using the entire register space, so in that respect, having one more register available doesn't help us much. see also the discussion we had previously [here](https://github.com/larq/compute-engine/issues/130).