@lgeiger 2020-04-14
Thanks for the links, makes sense. This probably only makes sense if we want to revisit an 8x8 BMLA or require more registers somehow.

> can you however benchmark it first?

I'll can run a quick benchmark later today.